.history/requirements_20251019212349.txt
torch
transformers
datasets
unsloth
trl
fastapi
uvicorn
mlflow
Pillow
pyyaml
requests
